Moved some of usermodel function into helper
