+++ lib/dailyploy/helper/user.ex
@@ -3,6 +3,8 @@ defmodule Dailyploy.Helper.User do
   alias Ecto.Multi
   alias Dailyploy.Schema.User
   alias Dailyploy.Schema.Company
+  alias Auth.Guardian
+  import Comeonin.Bcrypt, only: [checkpw: 2, dummy_checkpw: 0]
 
   alias Dailyploy.Repo
 
@@ -33,4 +35,41 @@ defmodule Dailyploy.Helper.User do
     |> Multi.insert(:user, user_changeset)
     |> Multi.insert(:company, company_changeset)
   end
+
+
+
+  def token_sign_in(email, password) do
+    case email_password_auth(email, password) do
+      {:ok, user} ->
+        Guardian.encode_and_sign(user)
+
+      _ ->
+        {:error, :unauthorized}
+    end
+  end
+
+  defp email_password_auth(email, password) when is_binary(email) and is_binary(password) do
+    with {:ok, user} <- get_by_email(email),
+         do: verify_password(password, user)
+  end
+
+  defp get_by_email(email) when is_binary(email) do
+    case Repo.get_by(User, email: email) do
+      nil ->
+        dummy_checkpw()
+        {:error, "Email does not match"}
+
+      user ->
+        {:ok, user}
+    end
+  end
+
+  defp verify_password(password, %User{} = user) when is_binary(password) do
+   if checkpw(password, user.password_hash) do
+      {:ok, user}
+   else
+     {:error, :invalid_password}
+   end
+  end
 end
+
