+++ lib/dailyploy/model/user.ex
@@ -1,10 +1,10 @@
 defmodule Dailyploy.Model.User do
   alias Dailyploy.Repo
 
-  alias Auth.Guardian
+
   alias Dailyploy.Schema.User
 
-  import Comeonin.Bcrypt, only: [checkpw: 2, dummy_checkpw: 0]
+
 
   @spec list_users :: any
   def list_users() do
@@ -27,39 +27,5 @@ defmodule Dailyploy.Model.User do
 
   def delete_user(%User{} = user) do
     Repo.delete(user)
-  end
-
-  def token_sign_in(email, password) do
-    case email_password_auth(email, password) do
-      {:ok, user} ->
-        Guardian.encode_and_sign(user)
-
-      _ ->
-        {:error, :unauthorized}
-    end
-  end
-
-  defp email_password_auth(email, password) when is_binary(email) and is_binary(password) do
-    with {:ok, user} <- get_by_email(email),
-         do: verify_password(password, user)
-  end
-
-  defp get_by_email(email) when is_binary(email) do
-    case Repo.get_by(User, email: email) do
-      nil ->
-        dummy_checkpw()
-        {:error, "Email does not match"}
-
-      user ->
-        {:ok, user}
-    end
-  end
-
-  defp verify_password(password, %User{} = user) when is_binary(password) do
-    if checkpw(password, user.password_hash) do
-      {:ok, user}
-    else
-      {:error, :invalid_password}
-    end
   end
 end
